Revert "Convert `create-slot` to TypeScript"
